Splits tests in two files and fixes README linting
